add:ログイン機能
従業員・管理者ログイン機能実装中

+++ app/database/migrations/2024_12_02_043918_create_masseges_table.php
@@ -14,7 +14,7 @@ class CreateMassegesTable extends Migration
     public function up()
     {
         Schema::create('masseges', function (Blueprint $table) {
-            $table->integer('id');
+            $table->bigIncrements('id');
             $table->integer('user_id');
             $table->string('body');
             $table->timestamps();
